Fix database script resolution for API deployments
